+++ src/provider/mod.rs
@@ -15,7 +15,6 @@ pub(crate) mod secp_secq;
 pub(crate) mod traits;
 // a non-hiding variant of {kzg, zeromorph}
 mod kzg_commitment;
-mod non_hiding_kzg;
 pub(crate) mod util;
 
 // crate-private modules
